@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <Project>
   <PropertyGroup>
-    <LangVersion>9.0</LangVersion>
+    <LangVersion>10.0</LangVersion>
     <Copyright>Pekka Heikura</Copyright>
     <PackageProjectUrl>https://github.com/pekkah/tanka-graphql</PackageProjectUrl>
     <RepositoryUrl>https://github.com/pekkah/tanka-graphql</RepositoryUrl>
