@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
     <AssemblyName>graphql.benchmarks</AssemblyName>
     <RootNamespace>Tanka.GraphQL.Benchmarks</RootNamespace>
     <IsPackable>false</IsPackable>
