@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
     <IsPackable>false</IsPackable>
     <AssemblyName>tanka.graphql.dev.allocations</AssemblyName>
     <RootNamespace>Tanka.GraphQL.Dev.Allocations</RootNamespace>
