@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFrameworks>netstandard2.1;net5.0</TargetFrameworks>
+    <TargetFrameworks>net6.0</TargetFrameworks>
     <RootNamespace>Tanka.GraphQL.Extensions.Tracing</RootNamespace>
     <AssemblyName>Tanka.GraphQL.Extensions.Tracing</AssemblyName>
   </PropertyGroup>
