@@ -1,15 +1,12 @@
-﻿
-using Tanka.GraphQL.Language.Nodes;
-using Tanka.GraphQL.TypeSystem;
+﻿using Tanka.GraphQL.Language.Nodes;
 
-namespace Tanka.GraphQL.Validation
+namespace Tanka.GraphQL.Validation;
+
+public struct VariableUsage
 {
-    public struct VariableUsage
-    {
-        public Variable Node;
+    public Variable Node;
 
-        public IType Type;
+    public TypeBase Type;
 
-        public object DefaultValue;
-    }
+    public object DefaultValue;
 }
