@@ -87,7 +87,7 @@ namespace Tanka.GraphQL.Benchmarks
         [GlobalSetup]
         public void Setup()
         {
-            _schema = Utils.InitializeSchema();
+            _schema = Utils.InitializeSchema().Result;
             _query = Utils.InitializeQuery();
             _mutation = Utils.InitializeMutation();
             _subscription = Utils.InitializeSubscription();
