@@ -1,6 +1,4 @@
-﻿using Tanka.GraphQL.SchemaBuilding;
-
-namespace Tanka.GraphQL.Directives
+﻿namespace Tanka.GraphQL.Directives
 {
     public delegate DirectiveVisitor CreateDirectiveVisitor(SchemaBuilder builder);
 }
