@@ -1,8 +1,8 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using Tanka.GraphQL.Language.Nodes;
+using Tanka.GraphQL.Language.Nodes.TypeSystem;
 using Tanka.GraphQL.TypeSystem;
-using Argument = Tanka.GraphQL.TypeSystem.Argument;
 
 namespace Tanka.GraphQL.Execution
 {
@@ -10,10 +10,10 @@ namespace Tanka.GraphQL.Execution
     {
         public static object? CoerceArgumentValue(
             ISchema schema,
-            IReadOnlyDictionary<string, object>? coercedVariableValues,
+            IReadOnlyDictionary<string, object?>? coercedVariableValues,
             string argumentName,
-            Argument argumentDefinition,
-            Language.Nodes.Argument argument)
+            InputValueDefinition argumentDefinition,
+            Argument? argument)
         {
             var argumentType = argumentDefinition.Type;
             var defaultValue = argumentDefinition.DefaultValue;
@@ -41,7 +41,7 @@ namespace Tanka.GraphQL.Execution
                 value = argumentValue;
             }
 
-            if (argumentType is NonNull && (!hasValue || value == null))
+            if (argumentType is NonNullType && (!hasValue || value == null))
                 throw new ValueCoercionException(
                     $"Argument '{argumentName}' is non-null but no value could be coerced",
                     null,
@@ -54,8 +54,7 @@ namespace Tanka.GraphQL.Execution
                 if (argumentValue is Variable) return value;
 
                 var coercedValue = Values.CoerceValue(
-                    schema.GetInputFields,
-                    schema.GetValueConverter,
+                    schema,
                     value,
                     argumentType);
 
@@ -67,24 +66,24 @@ namespace Tanka.GraphQL.Execution
 
         public static IReadOnlyDictionary<string, object?> CoerceArgumentValues(
             ISchema schema,
-            ObjectType objectType,
+            ObjectDefinition objectDefinition,
             FieldSelection field,
-            IReadOnlyDictionary<string, object> coercedVariableValues)
+            IReadOnlyDictionary<string, object?> coercedVariableValues)
         {
             var coercedValues = new Dictionary<string, object?>();
 
             var argumentValues = field.Arguments ?? Arguments.None;
             var fieldName = field.Name;
-            var argumentDefinitions = schema.GetField(objectType.Name, fieldName)
+            var argumentDefinitions = schema.GetRequiredField(objectDefinition.Name, fieldName)
                 .Arguments;
 
             if (argumentDefinitions == null)
                 return coercedValues;
 
-            foreach (var argumentDefinitionPair in argumentDefinitions)
+            foreach (var definition in argumentDefinitions)
             {
-                var argumentDefinition = argumentDefinitionPair.Value;
-                Name argumentName = argumentDefinitionPair.Key;
+                var argumentDefinition = definition;
+                Name argumentName = definition.Name;
                 var argument = argumentValues.SingleOrDefault(a => a.Name == argumentName);
                 coercedValues[argumentName] = CoerceArgumentValue(
                     schema,
