@@ -1,28 +1,29 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
+using Tanka.GraphQL.Language.Nodes;
 
 namespace Tanka.GraphQL.TypeSystem
 {
-    public class DirectiveList : IHasDirectives, IEnumerable<DirectiveInstance>
+    public class DirectiveList : IHasDirectives, IEnumerable<Directive>
     {
-        private readonly Dictionary<string, DirectiveInstance>
-            _directives = new Dictionary<string, DirectiveInstance>();
+        private readonly Dictionary<string, Directive>
+            _directives = new Dictionary<string, Directive>();
 
-        public DirectiveList(IEnumerable<DirectiveInstance>? directives = null)
+        public DirectiveList(IEnumerable<Directive>? directives = null)
         {
             if (directives != null)
                 foreach (var directiveInstance in directives)
                     _directives[directiveInstance.Name] = directiveInstance;
         }
 
-        public IEnumerable<DirectiveInstance> Directives => _directives.Values;
+        public IEnumerable<Directive> Directives => _directives.Values;
 
-        public DirectiveInstance? GetDirective(string name)
+        public Directive? GetDirective(string name)
         {
             return _directives.ContainsKey(name) ? _directives[name] : null;
         }
 
-        public IEnumerator<DirectiveInstance> GetEnumerator()
+        public IEnumerator<Directive> GetEnumerator()
         {
             return _directives.Values.GetEnumerator();
         }
