@@ -1,41 +1,44 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using Tanka.GraphQL.Language.Nodes.TypeSystem;
 using Tanka.GraphQL.TypeSystem.ValueSerialization;
 using Tanka.GraphQL.ValueResolution;
 
 namespace Tanka.GraphQL.TypeSystem
 {
-    public interface ISchema : IHasDirectives, IType
+    public interface ISchema : IHasDirectives
     {
-        ObjectType? Subscription { get; }
+        ObjectDefinition? Mutation { get; }
 
-        ObjectType Query { get; }
+        ObjectDefinition Query { get; }
 
-        ObjectType? Mutation { get; }
+        ObjectDefinition? Subscription { get; }
 
-        INamedType GetNamedType(string name);
+        TypeDefinition? GetNamedType(string name);
 
-        IField GetField(string type, string name);
+        FieldDefinition? GetField(string type, string name);
 
-        IEnumerable<KeyValuePair<string, IField>> GetFields(string type);
+        IEnumerable<KeyValuePair<string, FieldDefinition>> GetFields(string type);
 
-        IQueryable<T> QueryTypes<T>(Predicate<T>? filter = null) where T : INamedType;
+        IQueryable<T> QueryTypes<T>(Predicate<T>? filter = null) where T : TypeDefinition;
 
-        DirectiveType GetDirectiveType(string name);
+        DirectiveDefinition? GetDirectiveType(string name);
 
-        IQueryable<DirectiveType> QueryDirectiveTypes(Predicate<DirectiveType> filter = null);
+        IQueryable<DirectiveDefinition> QueryDirectiveTypes(Predicate<DirectiveDefinition>? filter = null);
 
-        IEnumerable<KeyValuePair<string, InputObjectField>> GetInputFields(string type);
+        IEnumerable<KeyValuePair<string, InputValueDefinition>> GetInputFields(string type);
 
-        InputObjectField GetInputField(string type, string name);
+        InputValueDefinition? GetInputField(string type, string name);
 
-        IEnumerable<ObjectType> GetPossibleTypes(IAbstractType abstractType);
+        IEnumerable<ObjectDefinition> GetPossibleTypes(InterfaceDefinition abstractType);
 
-        Resolver GetResolver(string type, string fieldName);
+        IEnumerable<ObjectDefinition> GetPossibleTypes(UnionDefinition abstractType);
 
-        Subscriber GetSubscriber(string type, string fieldName);
+        Resolver? GetResolver(string type, string fieldName);
 
-        IValueConverter GetValueConverter(string type);
+        Subscriber? GetSubscriber(string type, string fieldName);
+
+        IValueConverter? GetValueConverter(string type);
     }
 }
