@@ -1,15 +1,18 @@
 ﻿using System.Collections.Generic;
+using System.Diagnostics.CodeAnalysis;
+using Tanka.GraphQL.Language.Nodes.TypeSystem;
 
 namespace Tanka.GraphQL.TypeSystem
 {
     public static class TypeDictionaryExtensions
     {
-        public static bool TryGetValue<T>(this IDictionary<string, INamedType> types, string key, out T type)
-            where T : INamedType
+        public static bool TryGetValue<T>(this IDictionary<string, TypeDefinition> types, string key,
+            [NotNullWhen(true)] out T? type)
+            where T : TypeDefinition
         {
             if (types.TryGetValue(key, out var value))
             {
-                type = (T) value;
+                type = (T)value;
                 return true;
             }
 
