@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -6,13 +7,13 @@ namespace Tanka.GraphQL.Validation
 {
     public class ValidationResult
     {
-        public static ValidationResult Success => new ValidationResult();
+        public IEnumerable<ValidationError> Errors { get; set; } = Array.Empty<ValidationError>();
+
+        public IReadOnlyDictionary<string, object>? Extensions { get; set; }
 
         public bool IsValid => !Errors.Any();
 
-        public IEnumerable<ValidationError> Errors { get; set; } = new ValidationError[0];
-
-        public IReadOnlyDictionary<string, object> Extensions { get; set; }
+        public static ValidationResult Success => new();
 
         public override string ToString()
         {
@@ -22,10 +23,7 @@ namespace Tanka.GraphQL.Validation
             if (!IsValid)
             {
                 builder.AppendLine("ExecutionErrors:");
-                foreach (var validationError in Errors)
-                {
-                    builder.AppendLine(validationError.ToString());
-                }
+                foreach (var validationError in Errors) builder.AppendLine(validationError.ToString());
             }
 
 
