@@ -2,7 +2,7 @@
 using System.Threading;
 using System.Threading.Tasks;
 using Tanka.GraphQL.Channels;
-using Tanka.GraphQL.TypeSystem;
+using Tanka.GraphQL.Language.Nodes.TypeSystem;
 
 namespace Tanka.GraphQL.ValueResolution
 {
@@ -14,24 +14,24 @@ namespace Tanka.GraphQL.ValueResolution
             return new ValueTask<ISubscriberResult>(eventChannel.Subscribe(unsubscribe));
         }
 
-        public static ValueTask<IResolverResult> As(object result)
+        public static ValueTask<IResolverResult> As(object? result)
         {
-            return new ValueTask<IResolverResult>(new CompleteValueResult(result, null));
+            return new ValueTask<IResolverResult>(new CompleteValueResult(result));
         }
 
-        public static ValueTask<IResolverResult> As(string result)
+        public static ValueTask<IResolverResult> As(string? result)
         {
-            return new ValueTask<IResolverResult>(new CompleteValueResult((object)result, null));
+            return new ValueTask<IResolverResult>(new CompleteValueResult(result));
         }
 
-        public static ValueTask<IResolverResult> As(ObjectType type, object result)
+        public static ValueTask<IResolverResult> As(ObjectDefinition type, object? result)
         {
             return new ValueTask<IResolverResult>(new CompleteValueResult(result, type));
         }
 
-        public static ValueTask<IResolverResult> As(IEnumerable result)
+        public static ValueTask<IResolverResult> As(IEnumerable? result)
         {
-            return new ValueTask<IResolverResult>(new CompleteValueResult(result, null));
+            return new ValueTask<IResolverResult>(new CompleteValueResult(result));
         }
     }
 }
