@@ -3,6 +3,7 @@ using System.Collections.Generic;
 
 using Tanka.GraphQL.Execution;
 using Tanka.GraphQL.Language.Nodes;
+using Tanka.GraphQL.Language.Nodes.TypeSystem;
 using Tanka.GraphQL.TypeSystem;
 
 namespace Tanka.GraphQL.ValueResolution
@@ -10,16 +11,16 @@ namespace Tanka.GraphQL.ValueResolution
     public class ResolverContext : IResolverContext
     {
         public ResolverContext(
-            ObjectType objectType,
-            object objectValue,
-            IField field,
+            ObjectDefinition objectDefinition,
+            object? objectValue,
+            FieldDefinition field,
             FieldSelection selection,
             IReadOnlyCollection<FieldSelection> fields,
             IReadOnlyDictionary<string, object?> arguments,
             NodePath path,
             IExecutorContext executionContext)
         {
-            ObjectType = objectType ?? throw new ArgumentNullException(nameof(objectType));
+            ObjectDefinition = objectDefinition ?? throw new ArgumentNullException(nameof(objectDefinition));
             ObjectValue = objectValue;
             Field = field ?? throw new ArgumentNullException(nameof(field));
             Selection = selection ?? throw new ArgumentNullException(nameof(selection));
@@ -33,17 +34,17 @@ namespace Tanka.GraphQL.ValueResolution
 
         public ISchema Schema => ExecutionContext.Schema;
 
-        public ObjectType ObjectType { get; }
+        public ObjectDefinition ObjectDefinition { get; }
 
-        public object ObjectValue { get; }
+        public object? ObjectValue { get; }
 
-        public IField Field { get; }
+        public FieldDefinition Field { get; }
 
         public FieldSelection Selection { get; }
         
         public IReadOnlyCollection<FieldSelection> Fields { get; }
 
-        public IReadOnlyDictionary<string, object> Arguments { get; }
+        public IReadOnlyDictionary<string, object?> Arguments { get; }
 
         public NodePath Path { get; }
 
