@@ -1,18 +1,18 @@
 ﻿using System;
-using Tanka.GraphQL.TypeSystem;
+using Tanka.GraphQL.Language.Nodes;
 
 namespace Tanka.GraphQL
 {
     public class VariableException : Exception
     {
-        public VariableException(string message, string variableName, IType variableType): base(message)
+        public VariableException(string message, string variableName, INode variableType) : base(message)
         {
             VariableName = variableName;
             VariableType = variableType;
         }
 
-        public IType VariableType { get; set; }
-
         public string VariableName { get; set; }
+
+        public INode VariableType { get; set; }
     }
 }
