@@ -1,6 +1,4 @@
 ﻿using System.Threading.Tasks;
-using Tanka.GraphQL.SchemaBuilding;
-using Tanka.GraphQL.SDL;
 using Tanka.GraphQL.TypeSystem;
 using Xunit;
 
@@ -24,18 +22,20 @@ namespace Tanka.GraphQL.Tests.Language.ImportProviders
                  ";
 
             /* When */
-            var builder = await new SchemaBuilder()
+            var builder = new SchemaBuilder()
                 // BuiltIn import providers are used
-                .SdlAsync(sdl);
+                .Add(sdl);
 
-            var schema = builder.Build();
+            var schema = await builder.Build(new SchemaBuildOptions());
 
             /* Then */
-            var importedType = schema.GetNamedType<ObjectType>("ImportedType");
+            var importedType = schema.GetNamedType("ImportedType");
             Assert.NotNull(importedType);
+           
             var importedField = schema.GetField(schema.Query.Name, "imported");
-            Assert.Same(importedType, importedField.Type.Unwrap());
-            var nestedType = schema.GetNamedType<ObjectType>("NestedObject");
+            Assert.Equal(importedType.Name, importedField.Type.Unwrap().Name);
+            
+            var nestedType = schema.GetNamedType("NestedObject");
             Assert.NotNull(nestedType);
         }
     }
